Re-parent Route and Controller off the classic object model
Route extends FrameworkObject instead of
EmberObject.extend(ActionHandler, Evented): ActionHandler's send() is
implemented directly, activate/deactivate events go through an owned
EventedEmitter, the 200-line reopen is folded into the class body with
prototype-assigned defaults (mergedProperties reproduces what
ActionHandler and the reopen contributed, so classic .extend() keeps
merging actions/queryParams), and the @computed _store/_qp getters
become a plain getter and a per-instance memo respectively.

Controller's behavior moves into a native class in -base.ts: init,
send(), _qpChanged, and a model accessor whose setter skips notification
for already-cached values and is registered in COMPUTED_SETTERS (like
@tracked setters) so metal set takes the fast path — the DEBUG
mandatory-setter pre-read otherwise consumes the tag mid-write and trips
the backtracking-rerender assertion on {{mount}}'s model updates. The
classic ControllerMixin remains public API, defined in index.ts but no
longer applied to the class.
